+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Command deduplication: Avoid handling the same command multiple times
 - Process-manager lifecycle hardening (no zombies)
 - Surface read-model freshness to clients
+- Aligned Quarkus with Spring Boot implemenation
 
 ## 0.6.0
 - Added new Jackson module